Additional debugging for 1.6.8.
Solution for 2.1 (though I'm now using the online judge number scheme).

This is not submitted in C++ because this problem is a clear C problem. The bit-bucket like data tracking system and early validity failure are all streamlined for a sequential approach. I don't believe the use of complicated data structures or sets is useful here other than a preallocated existence set as I've done.