2.8.4 Solved!
The time limit exceeded problem was because of an error in my previous
code where I was not clearing the guesses from the previous state. Here
is the result from the UVA online judge:

Hi,

This is an automated response from UVa Online Judge.

Your submission with number 12568817 for the problem 843 - Crypt Kicker
has succeeded with verdict Accepted.

Congratulations! Now it is time to try a new problem.

Best regards,

The UVa Online Judge team