This is my solution to WordLadder2 after several iterations.
Online judge is not passed due to timeout. However, my local timings on a release build run one of the slowest problems in 94ms and then the following problem where I hit the next timeout in sub 16ms. May be a memory issue so I'm going to try and reduce the memory overhead of the problem next.