Ermagod, this finally passes the judge.
Note I had some bugs. My path rebuilder didn't take into account terminals of the same word break. For the rest of the performance I simply removed usage of string in a bunch of places and favored pointers. This must have eventually led to the VM being able to finish the program in the time allotted.